Rename CarePlan template
